schedule_flight success

+++ service_backend/src/main/java/com/cs4400/service_backend/mapper/FlightMapper.java
@@ -9,7 +9,7 @@ import java.sql.Time;
 @Mapper
 public interface FlightMapper {
     void schedule_flight(String flight_num, String airline_name, String  from_airport , String to_airport,
-                           Time departure_time, Time arrival_time, Date flight_date, int cost, int capacity);
+                           Time departure_time, Time arrival_time, Date flight_date, double cost, int capacity);
 
     Flight check_flight(String flight_num, String airline_name);
 }
